-added comments to all plants -changed sun class to use constant y velocity of about 1s per lawn tile -fixed lawnmower

+++ plant.cpp
@@ -16,7 +16,7 @@ void Plant::decreaseHealth(int value)
     life -= value;
 }
 
-int Plant::getDamage() const
+double Plant::getDamage() const
 {
     return plantDamage;
 }
